Support all handler in 1 lib file
