+++ plugin/test-event/src/mapping.rs
@@ -1,5 +1,5 @@
 use crate::models::EventTs;
-use massbit_chain_substrate::data_type::SubstrateEventRecord;
+use massbit_chain_substrate::data_type::{SubstrateEventRecord, SubstrateUncheckedExtrinsic, SubstrateBlock};
 use chrono;
 
 pub fn handle_event(event: &SubstrateEventRecord) -> Result<(), Box<dyn std::error::Error>> {
@@ -11,3 +11,13 @@ pub fn handle_event(event: &SubstrateEventRecord) -> Result<(), Box<dyn std::err
     event_ts.save();
     Ok(())
 }
+
+pub fn handle_extrinsic(extrinsic: &SubstrateUncheckedExtrinsic) -> Result<(), Box<dyn std::error::Error>> {
+    Ok(())
+}
+
+pub fn handle_block(block: &SubstrateBlock) -> Result<(), Box<dyn std::error::Error>> {
+    Ok(())
+}
+
+
