+++ plugin/test-extrinsic/src/mapping.rs
@@ -1,5 +1,5 @@
 use crate::models::ExtrinsicTs;
-use massbit_chain_substrate::data_type::SubstrateUncheckedExtrinsic;
+use massbit_chain_substrate::data_type::{SubstrateUncheckedExtrinsic, SubstrateEventRecord, SubstrateBlock};
 
 pub fn handle_extrinsic(extrinsic: &SubstrateUncheckedExtrinsic) -> Result<(), Box<dyn std::error::Error>> {
     println!("[SO File] Received Extrinsic");
@@ -9,5 +9,13 @@ pub fn handle_extrinsic(extrinsic: &SubstrateUncheckedExtrinsic) -> Result<(), B
         extrinsic: string_extrinsic,
     };
     extrinsic_ts.save();
+    Ok(())
+}
+
+pub fn handle_event(event: &SubstrateEventRecord) -> Result<(), Box<dyn std::error::Error>> {
+    Ok(())
+}
+
+pub fn handle_block(block: &SubstrateBlock) -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
